agents/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ag-ui-protocol>=0.1.9",
    "agno>=2.0.11",
    "fastapi[standard]>=0.117.1",
    "google-genai>=1.39.1",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
